pull tweet view into separate class
